backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic>=2.7.0
pydantic[email]>=2.7.0
pydantic-settings>=2.9.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.0.1
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2 